+++ drivers/console/CMakeLists.txt
@@ -13,5 +13,4 @@ zephyr_library_sources_ifdef(CONFIG_SEMIHOST_CONSOLE semihost_console.c)
 zephyr_library_sources_ifdef(CONFIG_UART_CONSOLE uart_console.c)
 zephyr_library_sources_ifdef(CONFIG_UART_MCUMGR uart_mcumgr.c)
 zephyr_library_sources_ifdef(CONFIG_UART_MUX uart_mux.c)
-zephyr_library_sources_ifdef(CONFIG_UART_PIPE uart_pipe.c)
 zephyr_library_sources_ifdef(CONFIG_XTENSA_SIM_CONSOLE xtensa_sim_console.c)
