+++ drivers/ieee802154/ieee802154_uart_pipe.c
@@ -21,7 +21,7 @@ LOG_MODULE_REGISTER(LOG_MODULE_NAME);
 #include <zephyr/net/net_pkt.h>
 #include <zephyr/random/rand32.h>
 
-#include <zephyr/drivers/console/uart_pipe.h>
+#include <zephyr/drivers/uart_pipe.h>
 #include <zephyr/net/ieee802154_radio.h>
 
 #include "ieee802154_uart_pipe.h"
