+++ drivers/serial/Kconfig
@@ -101,6 +101,27 @@ config UART_WIDE_DATA
 	  This is up to the driver to implement the necessary functions
 	  to properly support this.
 
+config UART_PIPE
+	bool "Pipe UART driver"
+	select UART_INTERRUPT_DRIVEN
+	help
+	  Enable pipe UART driver. This driver allows application to communicate
+	  over UART with custom defined protocol. Driver doesn't inspect received
+	  data (as contrary to console UART driver) and all aspects of received
+	  protocol data are handled by application provided callback.
+
+# Workaround for not being able to have commas in macro arguments
+DT_CHOSEN_Z_UART_PIPE := zephyr,uart-pipe
+
+config UART_PIPE_ON_DEV_NAME
+	string "Device Name of UART Device for pipe UART"
+	default "$(dt_chosen_label,$(DT_CHOSEN_Z_UART_PIPE))" if HAS_DTS
+	default "UART_0"
+	depends on UART_PIPE
+	help
+	  This option specifies the name of UART device to be used
+	  for pipe UART.
+
 comment "Serial Drivers"
 
 source "drivers/serial/Kconfig.b91"
