+++ drivers/serial/uart_pipe.c
@@ -12,13 +12,13 @@
  */
 
 #include <zephyr/logging/log.h>
-LOG_MODULE_REGISTER(uart_pipe, CONFIG_UART_CONSOLE_LOG_LEVEL);
+LOG_MODULE_REGISTER(uart_pipe);
 
 #include <zephyr/kernel.h>
 
 #include <zephyr/drivers/uart.h>
 
-#include <zephyr/drivers/console/uart_pipe.h>
+#include <zephyr/drivers/uart_pipe.h>
 #include <zephyr/sys/printk.h>
 
 static const struct device *uart_pipe_dev;
